Add files via upload
In this project, I used Python to perform a comprehensive analysis of the real estate market and uncover valuable insights. I cleaned and prepared the data, merged two different datasets, and explored various aspects of the market, such as customer age, property prices, and customer satisfaction. I also created various visualizations to display the data and identify the trends and patterns. I then created a profile of the ideal customer and the different types of buildings, which helped me to target and attract potential buyers more effectively. This project demonstrates my skills in data analysis, data visualization, and Python programming.